requirements.txt
pillow
pyautogui
pytesseract
